`mutectrl msg add --mail-input` (for Mutt)
Add option --mail-input to `mutectrl msg add` which parses the input as
an email message. The manadatory 'To' header field is used as recipient
and the optional 'Subject' header field combined with the email body as
the actual message.

This allows for a better integration with Mutt and MuteSend.sh is
changed accordingly.